Adopt commit a1c6b45 from master: Merge pull request #358 from hrosen4/sun-gcc-cast-align-drqs-81423822
commit a1c6b45e2eddbe551b22b6d6ce41b76e63be7d02
Merge: 4346f9b cb703ec
Author: Hyman Rosen <hrosen4@bloomberg.net>
Date:   Tue Apr 26 12:08:09 2016 -0400

    Merge pull request #358 from hrosen4/sun-gcc-cast-align-drqs-81423822

    Remove cast-align warnings with Sun gcc.

commit cb703ec178f730cfde8f8af2fe667459ea011f88
Author: Hyman Rosen <hrosen4@bloomberg.net>
Date:   Tue Apr 26 10:20:46 2016 -0400

    Use plain char for offsetting pointers.

commit bba6965c373851393963e809cfd9744c30d948fe
Author: Hyman Rosen <hrosen4@bloomberg.net>
Date:   Mon Apr 25 11:00:03 2016 -0400

    Remove cast-align warnings with Sun gcc.

+++ groups/bal/balb/balb_performancemonitor.cpp
@@ -1061,13 +1061,13 @@ int balb::PerformanceMonitor::Collector<bsls::Platform::OsUnix>
     stats->d_lstData[e_VIRTUAL_SIZE]  =
         static_cast<double>(status.pr_brksize) / (1024 * 1024);
 
-    cpuTimeU = bsls::TimeInterval(static_cast<double>(status.pr_utime.tv_sec),
-                                 static_cast<double>(status.pr_utime.tv_nsec)).
-                                                     totalSecondsAsDouble();
+    cpuTimeU =
+        bsls::TimeInterval(status.pr_utime.tv_sec, status.pr_utime.tv_nsec)
+            .totalSecondsAsDouble();
 
-    cpuTimeS = bsls::TimeInterval(static_cast<double>(status.pr_stime.tv_sec),
-                                 static_cast<double>(status.pr_stime.tv_nsec)).
-                                                     totalSecondsAsDouble();
+    cpuTimeS =
+        bsls::TimeInterval(status.pr_stime.tv_sec, status.pr_stime.tv_nsec)
+            .totalSecondsAsDouble();
 
     close(fd);
 #else
